requirements.txt
streamlit
faiss-cpu
sentence-transformers
pdfplumber
nltk
tqdm
python-dotenv
google-generativeai
